修正 refs #5922

+++ ticketing/src/altair/app/ticketing/memberships/views.py
@@ -12,7 +12,7 @@ from altair.app.ticketing.users import models as umodels
 from altair.app.ticketing.views import BaseView
 from altair.app.ticketing.fanstatic import with_bootstrap
 from . import forms
-
+from altair.app.ticketing.response import refresh_response
 ## admin権限を持っている人以外見れない想定。
 
 @view_defaults(permission="administrator", decorator=with_bootstrap, route_name="memberships")
@@ -78,14 +78,17 @@ class MembershipView(BaseView):
         self.request.session.flash(u"membershipを編集しました")
         return HTTPFound(self.request.route_url("memberships", action="index", membership_id="*"))
 
-    @view_config(match_param="action=delete", renderer="altair.app.ticketing:templates/memberships/delete.html")
+    @view_config(match_param="action=delete", renderer="altair.app.ticketing:templates/memberships/_delete_modal.html")
     def delete(self):
         membership = umodels.Membership.query.filter_by(id=self.request.matchdict["membership_id"]).first()
         if membership is None:
             raise HTTPNotFound
+        if len(membership.membergroups) > 0:
+            return {"message": u"{membership.name}には１つ以上の設定が存在しています。消せません。"}
         membership.delete()
         self.request.session.flash(u"membershipを削除しました")
-        return HTTPFound(self.request.route_url("memberships", action="index", membership_id="*"))
+        return refresh_response(self.request, {"redirect_to": self.request.route_url("memberships", action="index", membership_id="*")})
+
 
 @view_defaults(decorator=with_bootstrap, route_name="membergroups", permission="administrator")
 class MemberGroupView(BaseView):
